+++ server/controllers/foods.controller.js
@@ -3,11 +3,13 @@ const foodsService = require('services/foods.service');
 const validateRequest = require('middlewares/validate-request');
 
 const get = (req, res, next) => {
+  if (req.query.accountId !== req.user.id) {
+    return res.status(401).json({ message: 'Unauthorized' });
+  }
   foodsService.get(req.query).then(food => res.json(food)).catch(next);
 }
 
 const createSchema = (req, res, next) => {
-  console.log(req.body)
   const schema = Joi.object({
     accountId: Joi.string().required(),
     date: Joi.date().required(),
@@ -16,7 +18,7 @@ const createSchema = (req, res, next) => {
     servings: Joi.number().required(),
     label: Joi.string().required(),
     category: Joi.string().required(),
-    foodContentsLabel: Joi.array().items(Joi.string()),
+    healthLabels: Joi.array().items(Joi.string()),
     image: Joi.string(),
     nutrients: Joi.object().required(),
   });
@@ -24,6 +26,9 @@ const createSchema = (req, res, next) => {
 }
 
 const create = (req, res, next) => {
+  if (req.body.accountId !== req.user.id) {
+    return res.status(401).json({ message: 'Unauthorized' });
+  }
   foodsService.create(req.body).then(food => res.json(food)).catch(next);
 }
 
@@ -37,7 +42,7 @@ const updateSchema = (req, res, next) => {
     servings: Joi.number().required(),
     label: Joi.string().required(),
     category: Joi.string().required(),
-    foodContentsLabel: Joi.array().items(Joi.string()),
+    healthLabels: Joi.array().items(Joi.string()),
     image: Joi.string(),
     nutrients: Joi.object().required()
   });
@@ -45,11 +50,24 @@ const updateSchema = (req, res, next) => {
 }
 
 const update = (req, res, next) => {
+  if (req.body.accountId !== req.user.id) {
+    return res.status(401).json({ message: 'Unauthorized' });
+  }
   foodsService.update(req.params.id, req.body).then(food => res.json(food)).catch(next);
 }
 
 const _delete = (req, res, next) => {
+  if (req.query.accountId !== req.user.id) {
+    return res.status(401).json({ message: 'Unauthorized' });
+  }
   foodsService.delete(req.params.id).then(food => res.json(food)).catch(next);
+}
+
+const bulkDelete = (req, res, next) => {
+  if (req.query.accountId !== req.user.id) {
+    return res.status(401).json({ message: 'Unauthorized' });
+  }
+  foodsService.bulkDelete(req.query.accountId).then(food => res.json(food)).catch(next);
 }
 
 module.exports = {
@@ -58,5 +76,6 @@ module.exports = {
   createSchema,
   update,
   updateSchema,
-  delete: _delete
+  delete: _delete,
+  bulkDelete
 }
